Fixes so that johto mode still works

+++ RandomizerRom.py
@@ -272,37 +272,38 @@ def WriteSpecialWildLevels(locationDict,distDict,monFun):
 			entry = open("Special Pokemon Locations/"+file,'r')
 			yamlData = yaml.load(entry)
 			loc = yamlData["Location"]
-			fileName = locationDict[loc].FileName
-			mon = yamlData["NormalMon"]
-			shift = yamlData["LevelShift"]
-			type = yamlData["Type"]
-			code = yamlData["Code"]
-			#we don't bother with having restrictions on these, as in general these pokemon are potentially missable
-			newmon = monFun(mon,1001)
-			givestr = ""
-			if(type == 'Give Mon with Berry'):
-					givestr = 'givepoke '+newmon+', '+str(distDict[loc]+shift)+', BERRY'
-			if(type == 'Wild Pokemon'):
-				givestr = 'loadwildmon '+newmon+', '+str(distDict[loc]+shift)
-			if(type == 'Give Egg'):
-				givestr = 'giveegg '+newmon+', '+str(distDict[loc]+shift)
-			if(type == 'Give Poke'):
-				givestr = 'givepoke '+newmon+', '+str(distDict[loc]+shift)
-			newcode = code.replace("MONLINE",givestr).replace("MONNAME",newmon)
-			#switch spaces to tabs.....
-			newcode = newcode.replace("    ","\t")
+			if(loc in locationDict):
+				fileName = locationDict[loc].FileName
+				mon = yamlData["NormalMon"]
+				shift = yamlData["LevelShift"]
+				type = yamlData["Type"]
+				code = yamlData["Code"]
+				#we don't bother with having restrictions on these, as in general these pokemon are potentially missable
+				newmon = monFun(mon,1001)
+				givestr = ""
+				if(type == 'Give Mon with Berry'):
+						givestr = 'givepoke '+newmon+', '+str(distDict[loc]+shift)+', BERRY'
+				if(type == 'Wild Pokemon'):
+					givestr = 'loadwildmon '+newmon+', '+str(distDict[loc]+shift)
+				if(type == 'Give Egg'):
+					givestr = 'giveegg '+newmon+', '+str(distDict[loc]+shift)
+				if(type == 'Give Poke'):
+					givestr = 'givepoke '+newmon+', '+str(distDict[loc]+shift)
+				newcode = code.replace("MONLINE",givestr).replace("MONNAME",newmon)
+				#switch spaces to tabs.....
+				newcode = newcode.replace("    ","\t")
 
-			#find the code we need to replace
-			coderegexstr = re.escape(code.replace("    ","\t")).replace("MONLINE",".+").replace("MONNAME",".+")
-			file = open("RandomizerRom/maps/"+fileName)
-			filecode = file.read()
-			oldcode = re.findall(coderegexstr,filecode)[0]
-			newfile = filecode.replace(oldcode,newcode)
-			#write the new file into the files for the randomizer rom
-			newfilestream = open("RandomizerRom/maps/"+fileName,'w')
-			newfilestream.seek(0)
-			newfilestream.write(newfile)
-			newfilestream.truncate()
-			newfilestream.flush()
-			#os.fsync(newfilestream.fileno())
-			newfilestream.close()
+				#find the code we need to replace
+				coderegexstr = re.escape(code.replace("    ","\t")).replace("MONLINE",".+").replace("MONNAME",".+")
+				file = open("RandomizerRom/maps/"+fileName)
+				filecode = file.read()
+				oldcode = re.findall(coderegexstr,filecode)[0]
+				newfile = filecode.replace(oldcode,newcode)
+				#write the new file into the files for the randomizer rom
+				newfilestream = open("RandomizerRom/maps/"+fileName,'w')
+				newfilestream.seek(0)
+				newfilestream.write(newfile)
+				newfilestream.truncate()
+				newfilestream.flush()
+				#os.fsync(newfilestream.fileno())
+				newfilestream.close()
